SPR-5298 JSP tag for building URLs with URI templates
Created spring:url and spring:param tags that enhance the functionality of the JSTL c:url tag.  URI templates are supported in the url value attribute.  They are resolved against the params defined inside the url tag body.  Params that are unable to be applied as URI template are added to the query string.

git-svn-id: https://src.springframework.org/svn/spring-framework/trunk@288 50f2f4bb-b051-0410-bef5-90022cba6387